Restructured example
